move to sqlalchemy0.6 and BigIngeger support

+++ FitsStorage.py
@@ -16,8 +16,7 @@ import re
 
 from sqlalchemy import Table, Column, MetaData, ForeignKey
 from sqlalchemy import desc, func
-from sqlalchemy import Integer, String, Boolean, Text, DateTime, Time, Date, Numeric
-from sqlalchemy.databases.postgres import PGBigInteger
+from sqlalchemy import Integer, String, Boolean, Text, DateTime, Time, Date, Numeric, BigInteger
 
 from sqlalchemy.orm import relation, backref, join
 
@@ -440,7 +439,7 @@ class TapeWrite(Base):
   startdate = Column(DateTime(timezone=False))
   enddate = Column(DateTime(timezone=False))
   suceeded = Column(Boolean)
-  size = Column(Integer)
+  size = Column(BigInteger)
   beforestatus = Column(Text)
   afterstatus = Column(Text)
   hostname = Column(Text)
